scripts/requirements.txt
# Python dependencies for PostgreSQL + pgvector + AWS Bedrock integration

# Database
psycopg2-binary>=2.9.9
pgvector>=0.2.4

# AWS
boto3>=1.34.0
botocore>=1.34.0

# Utilities
python-dotenv>=1.0.0
